Move tightly-coupled methods into the same module
