+++ app/models/attendance_event.rb
@@ -1,5 +1,4 @@
 class AttendanceEvent < ActiveRecord::Base
-  include DateToSchoolYear
   include AssignToSchoolYear
   belongs_to :student
   belongs_to :school_year
