+++ app/models/discipline_incident.rb
@@ -1,5 +1,4 @@
 class DisciplineIncident < ActiveRecord::Base
-  include DateToSchoolYear
   include AssignToSchoolYear
   belongs_to :student
   belongs_to :school_year
