+++ app/models/intervention.rb
@@ -7,7 +7,6 @@ class Intervention < ActiveRecord::Base
   validates_presence_of :student_id, :intervention_type_id, :start_date
   validate :end_date_cannot_come_before_start_date
   delegate :name, to: :intervention_type
-  include DateToSchoolYear
   include AssignToSchoolYear
   before_save :assign_to_school_year
