+++ app/models/school_year.rb
@@ -4,7 +4,6 @@ class SchoolYear < ActiveRecord::Base
   has_many :student_assessments, -> (student) { extending FindByStudent }
   has_many :interventions, -> (student) { extending FindByStudent }
   validates_uniqueness_of :name, :start
-  extend DateToSchoolYear
 
   def self.in_between(school_year_1, school_year_2)
     where(start: (school_year_1.start)..(school_year_2.start)).order(:start).reverse
